# dist/swagger/swagger.json
{
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "BOS config admin API routes",
        "version": "1.0.0",
        "description": "Welcome to the reference documentation for Portofolio asset manager (PAM) REST API! </br></br>To learn about the common use cases and concept this REST APIs, check out our resource center  (<a>https://resourcecenter.fr.spinalcom.com/ressources-développeur</a>).</br></br>In addition to the API Reference, we also provide complete documentation for adminnistration API, that need to install and implement Spinalcore Digital Twin Operating System on their built environment.",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Spinalcom",
            "email": "developers@spinalcom.com"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/pam/create_digitaltwin": {
            "post": {
                "operationId": "CreateDigitalTwin",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "string"
                                        },
                                        {
                                            "properties": {
                                                "message": {}
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "DigitalTwin"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "folderPath": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "folderPath",
                                    "name"
                                ],
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}